tests: rename info to env
