+++ tests/test_factory_constructors.py
@@ -2,7 +2,7 @@
 import pytest
 import re
 
-import info  # noqa: F401
+import env  # noqa: F401
 
 from pybind11_tests import factory_constructors as m
 from pybind11_tests.factory_constructors import tag
@@ -420,7 +420,7 @@ def test_reallocations(capture, msg):
     """)
 
 
-@pytest.mark.skipif("info.PY2")
+@pytest.mark.skipif("env.PY2")
 def test_invalid_self():
     """Tests invocation of the pybind-registered base class with an invalid `self` argument.  You
     can only actually do this on Python 3: Python 2 raises an exception itself if you try."""
