+++ tests/test_kwargs_and_defaults.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 import pytest
 
-import info  # noqa: F401
+import env  # noqa: F401
 
 from pybind11_tests import kwargs_and_defaults as m
 
@@ -149,7 +149,7 @@ def test_keyword_only_args(msg):
     """
 
 
-@pytest.mark.xfail("info.PYPY and info.PY2", reason="PyPy2 doesn't seem to double count")
+@pytest.mark.xfail("env.PYPY and env.PY2", reason="PyPy2 doesn't seem to double count")
 def test_args_refcount():
     """Issue/PR #1216 - py::args elements get double-inc_ref()ed when combined with regular
     arguments"""
