+++ tests/test_numpy_dtypes.py
@@ -3,7 +3,7 @@ import re
 
 import pytest
 
-import info  # noqa: F401
+import env  # noqa: F401
 
 from pybind11_tests import numpy_dtypes as m
 
@@ -295,7 +295,7 @@ def test_register_dtype():
     assert 'dtype is already registered' in str(excinfo.value)
 
 
-@pytest.mark.xfail("info.PYPY")
+@pytest.mark.xfail("env.PYPY")
 def test_str_leak():
     from sys import getrefcount
     fmt = "f4"
